requirements.txt
pyyaml==6.0.2
line-bot-sdk==3.12.0
openai==1.42.0
boto3==1.35.8